refactor: Streamlitアプリケーションをマルチページ形式に移行
ユーザーエクスペリエンスとコードの保守性を向上させるため、アプリケーションを2つのページに分離しました。

- `src/main.py`：アプリケーションの概要と説明を提供するランディングページ。
- `src/pages/1_analysis.py`：データアップロード、パラメータ設定、結果表示のためのすべての対話型コンポーネントを含む分析ページ。

以前 `src/components/` にあったUIコンポーネントのコードは新しい分析ページに統合され、不要になった `src/components/` ディレクトリは削除されました。
